bug fix
Now we can save user settings.
Trick, do not use !== "undefined", that is strictly testing if the left
is equal to the string "undefined".
Use !== undefined

Also when user uninstall the app, the storage will be empty. Since we
are using an object to reference value in a concantenation fashion
(data.setting.workHours), we will have undefined value on data.setting,
and then exception throws out that we cannot reference workHours of
undefined.

http://stackoverflow.com/questions/20991989/check-for-duplicate-record-in-chrome-storage-extension-before-saving

Another bug: Notification only show up once.